+++ refena/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.0.0
 
+- **BREAKING**: To dispatch global actions using `BuildContext` or `Ref`, use `context.global.dispatch` or `ref.global.dispatch` instead of `context.dispatch` or `ref.dispatch`
 - **BREAKING**: Remove `debugLabel` from notifier constructors, override `customDebugLabel` instead
 - **BREAKING**: `ViewModelBuilder`: change parameters of `init` and `initBuild` from `(context, ref)` to `(context)`
 - **BREAKING**: `ViewModelBuilder`: change `placeholder`, `error` to `loadingBuilder`, `errorBuilder`